Refactoring src\main.c
Trello: https://trello.com/c/CgAdOFfe

- modularization of main in seperate files
- Correct call of cpp function in c (extern "c" --> mymain.cpp)
- Little safety improvement with vTaskDelete(NULL) after while(1) loop in threads
